+++ src/main/java/com/viktor/restcontrollers_js/service/RoleService.java
@@ -11,7 +11,7 @@ public interface RoleService {
 
     List<Role> getAllRoles();
 
-    Role findById(Long id);
+    Role findById(int id);
 
     void saveRole(Role role);
 }
