+++ src/main/java/com/viktor/restcontrollers_js/service/RoleServiceImpl.java
@@ -26,7 +26,7 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Role findById(Long id) {
+    public Role findById(int id) {
         return roleDao.findById(id);
     }
